Changed id of tasks to string format(Still UUID based) and deadline to date instead of calendar. Also included in this commit is a basic auth UI flow that will sign in users into the app.
Reason: For firestore, to use the POJO custom converter, the class related
to our model has to store data that can be interpreted by the firestore
converter. So it could not accept UUID and calendar as they are not POJO
but more complex type objects with abstract classes and what not.