Unit Tests for models are written and tested.
1) Written unit tests for Task, TaskList and TaskChange
2) Added Mockito dependency
3) Fixed minor bugs in TaskList class

Unit tests have been written for the three model classes with
expected behaviour. Most method names in test classes follow the format:
[method name][state of test][expected results].

Fixed minor bugs in TaskList class where there will always be calls
to populate() method whenever anyone calls get(), thus populating
the list continuously. Also the sort method now updates the list of
tasks to the sorted one. This is because we wish to update users
with the most recently made sorted changes. No point in trying to
keep the older list around and it makes it easier for any ViewAdapters
to get the list now with the updated singleton.