Updated database rules to allow only users and added a debug fragment.
Changes made this commit:
1) A debug fragment for debugging/testing.
2) Update of firebase firestore rules
3) Added multidex dependency
4) Added additional SHA-1 fingerprint for another computer access
5) Fixed minor bugs in Auth and Database classes

Now there is a way to get into a debug fragment that allows you
to test code without UI considerations. This is made due to the lack of
unit test support for firestore database and no local run of the
database currently exists and the database instances require android
context tools to be able to run properly. Thus the use of a debug page.

The firestore rules have now been updated to allow only users to access
their own database.

Also, if you wish to have access to the firestore database, it is
necessary for you to provide a SHA-1 fingerprint for access for the
debug build and download the latest google-services.json file.

Now the Auth class is able to sign users out of the database and the
Database class only allows Firebase users to write to database
while hiding the methods that write by path. This is so that users of
this database API can only write if there is a user available. Otherwise,
it is not available to the non-signed in users.