Updated database with listeners for event completion, migrated to androidx.
Changes made this commit:
1) Migrated whole project to androidx support library
2) Updated Database with listeners for events completion
3) Removed redundant imports and changed android.support to androidx
4) Added documentation to Database and Auth classes
5) Added DatabaseLogger for logging database messages
6) Changed XML layout files to use new androidx library.

As of this commit, the android support library has been migrated over
to the androidx library and we are to use the androidx library from now
on. If there are errors in building the project, check the gradle.properties
file and ensure that the following lines are inside.
-- android.enableJetifier=true android.useAndroidX=true

Now the database is fully functional with CRUD operations and
event listeners that inform about completion of the async events.
There is one listener with 4 method representing the operations. This
is for easy use so that we don't have to register multiple listeners.
Also, the database package is mostly documented now so take a read if
you are unsure of how to use it.

Also, a new logging class has been added to help aid in Database logging
to Logcat and simplify the workflow. That class is to remain package-
private and not to be used outside of the database package for fear
of leaks.

The reason to migrate to androidx library is for API level 28 and above
support and that the latest firebase components and artifacts also
use the latest androidx library. However, since some components still
do not migrate tho the androidx library, the gradle.properties file
needs to have the jetifier.