Changes made in this commit: 1) Added AdditionFragment as well as AddTask Menu item on MainFragment. 2) Implemented Task update functionalities to DetailFragment. 3) Expanded TaskChange.Builder "add" method to methods specific to the changed field. 4) Make minor changes to DebugFragment to account for the changes in (3). 5) Modified AndroidManifest.xml file to optimize interaction between on-screen keyboard and View items. 6) Added some methods to Task for easier handling of data for the Status field.
Users can now easily add a new task to their WhatsNext to-do list by clicking on the '+' menu item on the menu
bar at the top right hand corner of the MainFragment page. This will bring them to the AdditionFragment page
where they can specify all the information of their tasks before adding the task to their task list by clicking
on the "Add task" button.

Users can now update the information of an existing task directly from the DetailFragment page of that
particular task. This can be done by changing the text in the respective EditTexts (for Name, Category,
Description), changing the Deadline TextView (by picking a new Date from the DatePickerDialog that is inflated
when the Deadline TextView is clicked on) or changing the Status Button (which cycles through the four default
statuses on each click). Upon detecting a change, the "Save changes" button at the bottom of the DetailFragment
will be enabled and updates will be made to the Task via a TaskChange object when the "Save changes" button is
clicked.

Previously, the "add" method in TaskChange.Builder is used for updates to all of the five fields by passing in a
String tag that specifies the field to be changed and an Object argument that contains the newly updated value
for that field. This generic "add" method is now replaced by five specific update methods for each of the
following fields so that argument types of these methods are only specific to the updated field.

Corrected the "updateDummyTaskWithNewNameSuccessfullyUpdatesDatabase" method in the DebugFragment to support
testing for the new methods introduced after the changes in the TaskChange.Builder update methods.

Added "android:windowSoftInputMode="adjustPan">" to the <activity> tag of AndroidManifest to prevent keyboard
from pushing the "Save changes" button up when editing fields in the DetailFragment page.

Included two additional static methods in Task.java to get the int representation of Status from its String
representation and vice-versa as there is frequent use of such conversion when accessing user input and writing
it to the Task or accessing the raw int value from the Task and converting it into String format to be displayed
in EditText / TextViews