Changes made in this commit: 1) Added communication between controller and database. 2) Updated authentication process. 3) Solve the bug of screen returning to a non-existing state after pressing back button. 4) Added a feature of Task deadlines appearing red in colour when it is overdue.
MainFragment can now get all Task from Database for displaying in the RecyclerView. AdditionFragment now adds a new Task to Database directly when "Add Task" button is clicked on. DetailFragment now updates existing Task on Database or deletes Task from Database when the respective "Update" and "Delete" buttons are clicked on.

Implemented signing out process and sign out menu item so that users can now sign out or switch accounts. Sign-in flow activity is started only in MainActivity. When user signs out, the MainActivity's sign-in in flow activity will be called to prompt for user sign in.

Due to the calling of addToBackStack() method for the FragmentTransactions when DetailFragment or AdditionFragment transits to MainFragment, a bug is introduced in which users can retrace back to the DetailFragment page of a previously deleted task or to the AdditionFragment page of a task that is already added. Therefore, the popBackStack method has been added to these specific areas to erase the Fragment from its backstack and prevent users from retracing to those states.