Refactored Database methods to accept user's collection path.
Changed made this commit:
1) Refactored CRUD methods in database to accept additional arg.
2) Refactor changes propogated to places with database code.

The 4 basic CRUD methods in the Database class now hold accept
extra argument to decide which folder/collection of the user to perform
the operation on. No checks are done on the new string arguments and
thus is insecure as deletion of user data can happen on developers
side. Consider refactoring string arguments into a path builder in the
future for security and code checking sakes.