Changes to be committed: 1) Implemented a page for Completed Tasks 2) Implemented sorting options 3) Refactored sign out process 4) Added a TaskComparatorFactory class
Allows user to refer to Tasks previously marked as done just in case there were useful content stored in the
description.

A sort fragment was created to allow easier selection of the various fields used for sorting.

Sign out and re-signing in via the startOnActivityResult is not implemented using a OnSignOutListener interface
added to MainFragment. This is because the previous version of keeping track of sign out status in the
MainFragment's onDetach method relies on dependency to the MainFragment and also creates a bug where the sign
screen keep popping out whenever the MainFragment is removed or replaced from the Fragment Transaction.

The TaskComparatorFactory allows easier and more specific creation of the required Comparator by providing
specific local methods for each type of field being compared.