Added app bar with drawer to the app.
Changed made this commit:
1) Added app bar with navigation drawer into the app.
2) Removed sign out option from overflow menu items.
3) Refactored sorting from fragment into nested menu items.
4) Minor refactoring of code around codebase.

The app bar is now added with the navigation drawer. The main container
for layouts is still the fragment_container and all fragments to be
added are still to be done the same way with the container. The reason
for the nested layouts was so that views would render properly. The
relative layout wrapping around the coordinator layout was to ensure
that the app bar does not hide the status bar at the top and the bar
will appear when it is hidden from being scrolled. Also, from now on,
in all static pages that do not scroll, any view that wishes to be
anchored to the bottom is required to include
android:layout_marginBottom="?attr/actionBarSize in order for it to be
displayed correctly due to the app bar pushing the views downwards.
However an apparent bug as mentioned in bug 1 todo in mainfragment
appears when you scroll the recyclerview and click on items to display
the detail fragment. The coordinator layout provided by google does not
dynamically render views onto the screen which causes such issues.

The sign out option has been shifted from overflow menu into the nav
drawer for clarity sake of using the app.

The sort fragment has been deleted and replaced in favour of the
nested menu items for sorting the lists. However, now the app does
not sort the lists when it is retrieved by the database, it is one of
the to do mentioned in mainfragment.

Removed multiple code like signOutFlow methods and a few unused var
that are lying around.

Also added in personal artwork for the app. Please give constructive
feedback.