Fixed bug where scrolling main list leads to appbar disappearing.
Changed made this commit:
1) Fixed bug of app bar disappearing.
2) Cleared up some access modifiers and XML attributes

The bug as described in bug 1 in the previous commit to this has been
solved where the appbar will disappear as the user scrolls downwards,
and thus hiding the appbar in the main recycler view in main fragment.
The workaround was to introduce a new fragment class that subclasses
the android fragment's class where we expand the appbar in the
onCreateView method. This helps ensure code is reusable and any
static(non scrollable) pages in the future will subclass the new
AbstractStaticFragment class in order to inherit the behaviour.