Included in swipe gesture support on main fragment's recycler view.
Changes made this commit:
1) Added swipe gesture functionality to main fragment
2) Refactored some string and style resources

The ItemTouchHelper has been added to the main fragment's recycler
view. This helps in adding swipe functionality on the recycler view's
task holders. Adheres to material design where users can interact with
the views displayed on the screen in an intuitive way. Consider
refactoring out the item touch helper into a separate class to keep
the main fragment's responsibility cleaner. Also, the views are impl
using a foreground/background concept, where all the views are inflated
in one shot but the foreground covers the background and using the
ItemTouchHelper's Ututil, it ensures that only the foreground is
movable. This was done to prevent canvas/graphics code from being
written and cluttering up the main fragment's code quality. However,
this causes the background views to be tied to an XML file and not be
as flexible as if the graphics code was written.

Refactored out some string resources that were hardcoded in the XML
files and included in some styles for the item touch helper's background
views.