Major refactor applied to codebase and inclusion of features.
Changes made this commit:
1) Major refactoring done to codebase
2) Inclusion of patterns package
3) Swipe to delete and archive(done) feature
4) Inclusion of additional simple listener to database
5) Additional image assets for project
6) Change of min SDK version for build

The main fragment's confusing application of done and not done list
has now been changed to better reflect the structure of the app.
The done and not done each get their own fragments now with both
having exactly the same structure, view and behaviour except a few
slight differences like list used and menu options. As such, a general
AbstractScrollableTaskFragment was made so that any fragment that wishes
to show a list of tasks that can scroll can subclass from the said
class and implement the abstract methods. Also, TaskList singleton's
method names have been changed to better reflect what it is doing
behind the scenes in the singleton. There are no arbitrary shifting of
tasks from lists to lists except when documented properly. Additionally,
multiple inner classes have been refactored out to make it easier to
edit and access.

Included in this commit is a patterns package modelling after software
engineering patterns. Only the Command pattern exists in the package
currently. The usage of the command pattern is done in favour of
multiple methods that cannot be reused with children classes. The
invoker can be passed around and with the correct command names can
execute commands as required.

List items are now swippable with left being delete and right being
archived(done). Also a simple database listener has been included
that does nothing in implementation of methods. This is simply to
reduce the amount of code being written.

The min. SDK version now used is API level 21 due to multiple context
resource dependent operations being flagged out by lint. There are no
pressures to support lower level API too so a system wide change is
a-okay.