Included in splash screen for loading of tasks from database.
Changes made this commit:
1) Splash screen shown to user for pre-loading of tasks
2) Changed database listener to give information on type of list

In order for the user to not experience any lag during the usage of
the app, a splash screen has been included to help fetch all the tasks
for the user from the database before loading up the main view for
the app. This enables smooth usage when user switches between the
different screens for the app.

In correspondence to the above, the database wrapper class now informs
all attached listeners of the type of list that has been queried
from the database so that it is easier to distinguish the type of
list that has been retrieved from the database.