Included FAB on main screen and solved minor bugs.
Changes made this commit:
1) Added in a FAB on the main screen to add tasks
2) Solved minor bugs
3) Included fade animations for multiple fragment transactions

Added a FAB adhering to Google's material design guidelines for
main action of the app. Removed the add icon from menu items to prevent
confusion of the user of having multiple ways to add a task.

Tasks which have their status change are now properly reflected in the
task done and not done list, alongside the corresponding database
actions to take. Also, the isOverdue method now properly returns
if a date is overdue. Bug was in the usage of after method.