Redesigned buttons in Addition and Detail fragments as well as replacing the Status toggle Buttons in these fragments to Spinner objects.
Changes made in this commit:
1) Replaced Status toggle Button with Spinner
2) Redesigned the utility buttons at the bottom of Addition and Detail Fragment

Status in Addition and Detail Fragments are now changed using a dropdown selector called a Spinner object. This
provides a more efficient and convenient selection process as compared to the previous toggle Button.

Added a Cancel Button in the Addition Fragment page to match the dual-button concept in Detail Fragment.
Improved the appearance of the Buttons to make it less blocky and greyish.