Made  changes to resource files to optimize view objects for Samsung S8.
Changes made in this commit:
1) Edited the text size and placement of View objects in various
   fragments
2) Minor bug fix in AbstractScrollableTaskFragment with regards to
   ItemTouchHelper

The app is ready for deployment. In fact, it has already been deployed
on Gary's Samsung S8, with all the widths and text sizes optimized for
Gary's viewing comfort. It has been a great pleasure and fruitful
learning experience working on this project. Summer 2019 wasn't such
a waste of time after all