Slight updates to Addition and Detail Fragment layout.
Changes made in this commit:
1) Small adjustments to Addition and Detail Fragment.

EditTexts in DetailFragment are now spaced by a guideline in the
ConstraintLayout. Since the Guideline follows a percentage on the
screen, it allows the view to be optimized not just for Samsung S8
but also for any other android devices. Update and Delete Button
have switched places as it is more natural to press on the right
button to Update like how the Add button is on the right of Addition
fragment.